refactor(base): mover módulos sem responsabilidade direta para o app djangosige
Foram movidos do app base para o app djangosige os módulos que não estavam alinhados com a responsabilidade principal do app base, a fim de melhorar a coesão e a organização do projeto.
