--- djangosige/base/context_version.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from djangosige import __version__
-
-
-def sige_version(request):
-    return {"versao": __version__}
