+++ djangosige/cadastro/views/produto.py
@@ -5,13 +5,9 @@ from datetime import datetime
 from django.db.models import F
 from django.urls import reverse_lazy
 
-from djangosige.base.custom_views import (
-    CustomCreateView,
-    CustomListView,
-    CustomUpdateView,
-)
 from djangosige.cadastro.forms import CategoriaForm, MarcaForm, ProdutoForm, UnidadeForm
 from djangosige.cadastro.models import Categoria, Fornecedor, Marca, Produto, Unidade
+from djangosige.custom_views import CustomCreateView, CustomListView, CustomUpdateView
 from djangosige.estoque.models import EntradaEstoque, ItensMovimento, ProdutoEstocado
