+++ djangosige/custom_views.py
@@ -5,10 +5,7 @@ from django.views.generic import ListView, TemplateView, View
 from django.views.generic.detail import DetailView
 from django.views.generic.edit import CreateView, UpdateView
 
-from djangosige.base.views_mixins import (
-    CheckPermissionMixin,
-    FormValidationMessageMixin,
-)
+from djangosige.views_mixins import CheckPermissionMixin, FormValidationMessageMixin
 
 
 class CustomView(CheckPermissionMixin, View):
