+++ djangosige/financeiro/views/plano.py
@@ -3,7 +3,7 @@
 from django.shortcuts import redirect
 from django.urls import reverse_lazy
 
-from djangosige.base.custom_views import (
+from djangosige.custom_views import (
     CustomCreateView,
     CustomTemplateView,
     CustomUpdateView,
