+++ djangosige/fiscal/views/nota_fiscal.py
@@ -5,15 +5,14 @@ from django.http import HttpResponse
 from django.shortcuts import redirect
 from django.urls import reverse_lazy
 
-from djangosige.base.custom_views import (
+from djangosige.cadastro.models import MinhaEmpresa
+from djangosige.custom_views import (
     CustomCreateView,
     CustomListView,
     CustomTemplateView,
     CustomUpdateView,
     CustomView,
 )
-from djangosige.base.views_mixins import FormValidationMessageMixin
-from djangosige.cadastro.models import MinhaEmpresa
 from djangosige.fiscal.forms import (
     AutXMLFormSet,
     BaixarNotaForm,
@@ -38,6 +37,7 @@ from djangosige.fiscal.models import (
 )
 from djangosige.login.models import Usuario
 from djangosige.vendas.models import ItensVenda, PedidoVenda
+from djangosige.views_mixins import FormValidationMessageMixin
 
 try:
     from .processador_nf import ProcessadorNotaFiscal
