+++ djangosige/vendas/views/vendas.py
@@ -9,13 +9,13 @@ from django.shortcuts import redirect
 from django.urls import reverse_lazy
 from geraldo.generators import PDFGenerator
 
-from djangosige.base.custom_views import (
+from djangosige.cadastro.models import MinhaEmpresa
+from djangosige.custom_views import (
     CustomCreateView,
     CustomListView,
     CustomUpdateView,
     CustomView,
 )
-from djangosige.cadastro.models import MinhaEmpresa
 from djangosige.login.models import Usuario
 from djangosige.vendas.forms import (
     ItensVendaFormSet,
